Unify name of test classes, organize them into packages, cleanup imports
