+++ test/macaroni/model/character/SaboteurTester.java
@@ -1,15 +1,13 @@
+package macaroni.model.character;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import macaroni.model.element.Pump;
 import macaroni.model.element.Pipe;
-import macaroni.model.character.Saboteur;
-
 
 import static org.mockito.Mockito.*;
-
-import macaroni.model.element.Element;
 
 public class SaboteurTester {
     private Pump mockPump;
