+++ test/macaroni/model/effects/BananaEffectTester.java
@@ -1,3 +1,5 @@
+package macaroni.model.effects;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import org.junit.jupiter.api.BeforeEach;
@@ -6,7 +8,6 @@ import org.junit.jupiter.api.Test;
 import macaroni.model.element.Pipe;
 import macaroni.model.element.Pump;
 import macaroni.model.character.Plumber;
-import macaroni.model.effects.BananaEffect;
 
 import static org.mockito.Mockito.*;
