+++ test/macaroni/model/effects/NoEffectTester.java
@@ -1,11 +1,11 @@
+package macaroni.model.effects;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-
 import macaroni.model.element.Pipe;
 import macaroni.model.character.Plumber;
-import macaroni.model.effects.NoEffect;
 
 
 import static org.mockito.Mockito.*;
