+++ test/macaroni/model/effects/TechnocolEffectTester.java
@@ -1,3 +1,5 @@
+package macaroni.model.effects;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import org.junit.jupiter.api.BeforeEach;
@@ -5,7 +7,6 @@ import org.junit.jupiter.api.Test;
 
 import macaroni.model.element.Pipe;
 import macaroni.model.character.Plumber;
-import macaroni.model.effects.TechnokolEffect;
 
 import static org.mockito.Mockito.*;
