+++ test/macaroni/model/misc/WaterCollectorTester.java
@@ -1,10 +1,10 @@
+package macaroni.model.misc;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import macaroni.model.misc.WaterCollector;
-
-public class WaterCollectorTest {
+public class WaterCollectorTester {
 
     private WaterCollector testWaterCollector;
