--- test/macaroni/test/SampleTest.java
@@ -1,14 +1,0 @@
-package macaroni.test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import org.junit.jupiter.api.Test;
-
-class SampleTest {
-
-	@Test
-	void test() {
-		assertEquals(0, 0);
-	}
-
-}
